🚚 [RUM] move errorCollection in its folder
